Cleaned up benchmark library
