+++ factrs-bench/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod tiny_solver;
+// mod sophus;
